Move format support into its own io module
