+++ src/codecs/gif.rs
@@ -36,17 +36,13 @@ use gif::{DisposalMethod, Frame};
 
 use crate::animation::{self, Ratio};
 use crate::color::{ColorType, Rgba};
-use crate::error::LimitError;
-use crate::error::LimitErrorKind;
 use crate::error::{
-    DecodingError, EncodingError, ImageError, ImageResult, ParameterError, ParameterErrorKind,
-    UnsupportedError, UnsupportedErrorKind,
+    DecodingError, EncodingError, ImageError, ImageResult, LimitError, LimitErrorKind,
+    ParameterError, ParameterErrorKind, UnsupportedError, UnsupportedErrorKind,
 };
-use crate::image::{AnimationDecoder, ImageDecoder, ImageFormat};
+use crate::image::{AnimationDecoder, ImageDecoder};
 use crate::traits::Pixel;
-use crate::ExtendedColorType;
-use crate::ImageBuffer;
-use crate::Limits;
+use crate::{ExtendedColorType, ImageBuffer, ImageFormat, Limits};
 
 /// GIF decoder
 pub struct GifDecoder<R: Read> {
