+++ src/codecs/ico/decoder.rs
@@ -6,7 +6,8 @@ use crate::color::ColorType;
 use crate::error::{
     DecodingError, ImageError, ImageResult, UnsupportedError, UnsupportedErrorKind,
 };
-use crate::image::{ImageDecoder, ImageFormat};
+use crate::image::ImageDecoder;
+use crate::ImageFormat;
 
 use self::InnerDecoder::*;
 use crate::codecs::bmp::BmpDecoder;
