+++ src/codecs/tga/decoder.rs
@@ -4,7 +4,8 @@ use crate::{
     error::{
         ImageError, ImageResult, LimitError, LimitErrorKind, UnsupportedError, UnsupportedErrorKind,
     },
-    image::{ImageDecoder, ImageFormat},
+    image::ImageDecoder,
+    ImageFormat,
 };
 use byteorder_lite::ReadBytesExt;
 use std::io::{self, Read};
