+++ src/codecs/tiff.rs
@@ -17,8 +17,9 @@ use crate::error::{
     DecodingError, EncodingError, ImageError, ImageResult, LimitError, LimitErrorKind,
     ParameterError, ParameterErrorKind, UnsupportedError, UnsupportedErrorKind,
 };
-use crate::image::{ImageDecoder, ImageEncoder, ImageFormat};
+use crate::image::{ImageDecoder, ImageEncoder};
 use crate::metadata::Orientation;
+use crate::ImageFormat;
 
 /// Decoder for TIFF images.
 pub struct TiffDecoder<R>
