+++ src/codecs/webp/decoder.rs
@@ -2,9 +2,11 @@ use std::io::{BufRead, Read, Seek};
 
 use crate::buffer::ConvertBuffer;
 use crate::error::{DecodingError, ImageError, ImageResult};
-use crate::image::{ImageDecoder, ImageFormat};
+use crate::image::ImageDecoder;
 use crate::metadata::Orientation;
-use crate::{AnimationDecoder, ColorType, Delay, Frame, Frames, RgbImage, Rgba, RgbaImage};
+use crate::{
+    AnimationDecoder, ColorType, Delay, Frame, Frames, ImageFormat, RgbImage, Rgba, RgbaImage,
+};
 
 /// WebP Image format decoder.
 ///
