+++ src/images/buffer.rs
@@ -9,12 +9,12 @@ use std::slice::{ChunksExact, ChunksExactMut};
 use crate::color::{FromColor, Luma, LumaA, Rgb, Rgba};
 use crate::error::ImageResult;
 use crate::flat::{FlatSamples, SampleLayout};
-use crate::image::{GenericImage, GenericImageView, ImageEncoder, ImageFormat};
+use crate::image::{GenericImage, GenericImageView, ImageEncoder};
 use crate::math::Rect;
 use crate::traits::{EncodableLayout, Pixel, PixelWithColorType};
 use crate::utils::expand_packed;
-use crate::DynamicImage;
 use crate::{save_buffer, save_buffer_with_format, write_buffer_with_format};
+use crate::{DynamicImage, ImageFormat};
 
 /// Iterate over pixel refs.
 pub struct Pixels<'a, P: Pixel + 'a>
