+++ src/io/free_functions.rs
@@ -9,10 +9,9 @@ use crate::error::{
     ImageError, ImageFormatHint, ImageResult, LimitError, LimitErrorKind, ParameterError,
     ParameterErrorKind, UnsupportedError, UnsupportedErrorKind,
 };
-use crate::image::ImageFormat;
 #[allow(unused_imports)] // When no features are supported
 use crate::image::{ImageDecoder, ImageEncoder};
-use crate::DynamicImage;
+use crate::{DynamicImage, ImageFormat};
 
 /// Create a new image from a Reader.
 ///
